# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Crew Challenges API
  version: 1.0.0
paths:
  /api/v1/auth/csrf:
    get:
      operationId: auth_csrf
      description: GET /api/v1/auth/csrf - sets the crew_csrftoken cookie; call it
        before the first POST.
      tags:
      - auth
      security:
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CsrfOut'
          description: ''
  /api/v1/auth/login:
    post:
      operationId: auth_login
      description: POST /api/v1/auth/login - starts a session. Then call GET /api/v1/me.
      tags:
      - auth
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginInRequest'
        required: true
      security:
      - cookieAuth: []
      - {}
      responses:
        '204':
          description: No response body
  /api/v1/auth/logout:
    post:
      operationId: auth_logout
      description: POST /api/v1/auth/logout - ends the session.
      tags:
      - auth
      security:
      - cookieAuth: []
      responses:
        '204':
          description: No response body
  /api/v1/challenges:
    get:
      operationId: challenges_list
      description: Scheduled challenges, by start date.
      parameters:
      - in: query
        name: phase
        schema:
          type: string
        description: 'Comma-separated: upcoming, active, finished. Default: all three.'
      tags:
      - challenges
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ChallengeOut'
          description: ''
    post:
      operationId: challenges_create
      description: Add a proposal to the crew's pool (409 pool_full when it is full).
      tags:
      - challenges
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChallengeInRequest'
        required: true
      security:
      - cookieAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChallengeOut'
          description: ''
  /api/v1/challenges/{challenge_id}:
    get:
      operationId: challenges_retrieve
      parameters:
      - in: path
        name: challenge_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - challenges
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChallengeOut'
          description: ''
    put:
      operationId: challenges_update
      description: The creator replaces their proposal while it is in the pool. Resets
        its votes.
      parameters:
      - in: path
        name: challenge_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - challenges
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChallengeInRequest'
        required: true
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChallengeOut'
          description: ''
    delete:
      operationId: challenges_destroy
      description: The creator or an admin withdraws a proposal from the pool.
      parameters:
      - in: path
        name: challenge_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - challenges
      security:
      - cookieAuth: []
      responses:
        '204':
          description: No response body
...
  /api/v1/challenges/{challenge_id}/participants:
    put:
      operationId: challenges_participants
      description: The creator changes who takes part while the challenge is a proposal.
      parameters:
      - in: path
        name: challenge_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - challenges
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ParticipantsInRequest'
        required: true
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChallengeOut'
          description: ''
  /api/v1/challenges/{challenge_id}/participation:
    delete:
      operationId: challenges_leave
      description: |-
        Opt out before the start, or leave a running challenge (today still counts).

        Either way the challenge is gone for the member afterwards (an admin still sees it).
      parameters:
      - in: path
        name: challenge_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - challenges
      security:
      - cookieAuth: []
      responses:
        '204':
          description: No response body
  /api/v1/challenges/{challenge_id}/schedule:
    put:
      operationId: challenges_schedule
      description: 'Admin: schedule a proposal for a period, or move it before it
        starts.'
      parameters:
      - in: path
        name: challenge_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - challenges
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScheduleInRequest'
        required: true
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChallengeOut'
          description: ''
    delete:
      operationId: challenges_unschedule
      description: 'Admin: put a scheduled challenge back in the pool before it starts.'
      parameters:
      - in: path
        name: challenge_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - challenges
      security:
      - cookieAuth: []
      responses:
        '204':
          description: No response body
  /api/v1/challenges/{challenge_id}/vote:
    put:
      operationId: challenges_vote
      description: Vote for a proposal (voting twice counts once).
      parameters:
      - in: path
        name: challenge_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - challenges
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChallengeOut'
          description: ''
    delete:
      operationId: challenges_vote_clear
      description: Take your vote back.
      parameters:
      - in: path
        name: challenge_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - challenges
      security:
      - cookieAuth: []
      responses:
        '204':
          description: No response body
...
  /api/v1/proposals:
    get:
      operationId: proposals_list
      description: The crew's pool of proposals, newest first, with votes and how
        full it is.
      tags:
      - proposals
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PoolOut'
          description: ''
...
components:
  schemas:
...
    ChallengeInRequest:
      type: object
      description: The creator's choices. Combinations are checked in services.clean_shape.
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 60
        rules:
          type: string
          default: ''
          maxLength: 500
        icon:
          allOf:
          - $ref: '#/components/schemas/IconEnum'
          default: star
        measure:
          $ref: '#/components/schemas/MeasureEnum'
        unit:
          type: string
          default: ''
          maxLength: 20
        frequency:
          $ref: '#/components/schemas/FrequencyEnum'
        weekdays:
          type: array
          items:
            type: integer
            maximum: 6
            minimum: 0
          description: 'For frequency=weekdays: 0 = Monday ... 6 = Sunday.'
          maxItems: 7
        times:
          type: integer
          minimum: 1
          nullable: true
        target_scope:
          allOf:
          - $ref: '#/components/schemas/TargetScopeEnum'
          default: none
        target_value:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
          nullable: true
        proof_kind:
          allOf:
          - $ref: '#/components/schemas/ProofKindEnum'
          default: none
        proof_required:
          type: boolean
          default: false
        participant_ids:
          type: array
          items:
            type: string
            format: uuid
          nullable: true
          description: 'Who takes part (the creator is always in). Create: default
            the whole crew. Edit: leave out to keep the list.'
      required:
      - frequency
      - measure
      - title
    ChallengeOut:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        rules:
          type: string
        icon:
          $ref: '#/components/schemas/IconEnum'
        measure:
          $ref: '#/components/schemas/MeasureEnum'
        unit:
          type: string
        frequency:
          $ref: '#/components/schemas/FrequencyEnum'
        weekdays:
          type: array
          items:
            type: integer
        times:
          type: integer
          nullable: true
        target_scope:
          $ref: '#/components/schemas/TargetScopeEnum'
        target_value:
          type: number
          format: double
          maximum: 100000000
          minimum: -100000000
          exclusiveMaximum: true
          exclusiveMinimum: true
          nullable: true
        proof_kind:
          $ref: '#/components/schemas/ProofKindEnum'
        proof_required:
          type: boolean
        state:
          $ref: '#/components/schemas/ChallengeStateEnum'
        phase:
          nullable: true
          oneOf:
          - $ref: '#/components/schemas/PhaseEnum'
          - $ref: '#/components/schemas/NullEnum'
        period_kind:
          nullable: true
          oneOf:
          - $ref: '#/components/schemas/PeriodKindEnum'
          - $ref: '#/components/schemas/NullEnum'
        period_start:
          type: string
          format: date
          nullable: true
        start_date:
          type: string
          format: date
          nullable: true
        end_date:
          type: string
          format: date
          nullable: true
        chosen_by:
          allOf:
          - $ref: '#/components/schemas/PersonOut'
          nullable: true
        chosen_at:
          type: string
          format: date-time
          nullable: true
        created_by:
          allOf:
          - $ref: '#/components/schemas/PersonOut'
          nullable: true
        created_at:
          type: string
          format: date-time
        revision:
          type: integer
        vote_count:
          type: integer
        voters:
          type: array
          items:
            $ref: '#/components/schemas/PersonOut'
        my_vote:
          type: boolean
          description: The current member voted for it.
        participants:
          type: array
          items:
            $ref: '#/components/schemas/ParticipantOut'
          description: Who takes part (and who left), in the order they joined the
            crew.
        taking_part:
          type: boolean
          description: The current member is a participant and has not left.
        mine:
          type: boolean
          description: Proposed by the current member.
      required:
      - chosen_at
      - chosen_by
      - created_at
      - created_by
      - end_date
      - frequency
      - icon
      - id
      - measure
      - mine
      - my_vote
      - participants
      - period_kind
      - period_start
      - phase
      - proof_kind
      - proof_required
      - revision
      - rules
      - start_date
      - state
      - taking_part
      - target_scope
      - target_value
      - times
      - title
      - unit
      - vote_count
      - voters
      - weekdays
    ChallengeStateEnum:
      enum:
      - proposed
      - chosen
      type: string
      description: |-
        * `proposed` - Proposed
        * `chosen` - Chosen
...
    CsrfOut:
      type: object
      properties:
        csrf_token:
          type: string
      required:
      - csrf_token
...
    FrequencyEnum:
      enum:
      - daily
      - weekdays
      - times_per_week
      - times_per_period
      - once
      type: string
      description: |-
        * `daily` - Every day
        * `weekdays` - Chosen days of the week
        * `times_per_week` - A number of times a week
        * `times_per_period` - A number of times in the period
        * `once` - Once, by the end
    IconEnum:
      enum:
      - dumbbell
      - running
      - book
      - water
      - sugar
      - phone
      - sleep
      - walk
      - meditate
      - food
      - money
      - star
      type: string
      description: |-
        * `dumbbell` - dumbbell
        * `running` - running
        * `book` - book
        * `water` - water
        * `sugar` - sugar
        * `phone` - phone
        * `sleep` - sleep
        * `walk` - walk
        * `meditate` - meditate
        * `food` - food
        * `money` - money
        * `star` - star
...
    LoginInRequest:
      type: object
      properties:
        username:
          type: string
          minLength: 1
          maxLength: 150
        password:
          type: string
          writeOnly: true
          minLength: 1
          maxLength: 128
      required:
      - password
      - username
...
    MeasureEnum:
      enum:
      - check
      - quantity
      - abstain
      type: string
      description: |-
        * `check` - Just check in
        * `quantity` - A number with a unit
        * `abstain` - Held it (did not do something)
...
    NullEnum:
      enum:
      - null
...
    ParticipantOut:
      type: object
      properties:
        member:
          $ref: '#/components/schemas/PersonOut'
        left_on:
          type: string
          format: date
          nullable: true
          description: Last day that counted, when they left during the challenge.
      required:
      - left_on
      - member
    ParticipantsInRequest:
      type: object
      properties:
        participant_ids:
          type: array
          items:
            type: string
            format: uuid
          description: Who takes part; the creator is always in.
      required:
      - participant_ids
...
    PeriodKindEnum:
      enum:
      - month
      - week
      - custom
      type: string
      description: |-
        * `month` - Month
        * `week` - Week
        * `custom` - Custom
    PersonOut:
      type: object
      properties:
        id:
          type: string
          format: uuid
        display_name:
          type: string
        avatar_seed:
          type: string
      required:
      - avatar_seed
      - display_name
      - id
    PhaseEnum:
      enum:
      - upcoming
      - active
      - finished
      type: string
      description: |-
        * `upcoming` - upcoming
        * `active` - active
        * `finished` - finished
    PoolOut:
      type: object
      properties:
        proposals:
          type: array
          items:
            $ref: '#/components/schemas/ChallengeOut'
          description: Newest first.
        size:
          type: integer
          description: Proposals in the pool.
        limit:
          type: integer
          description: How many the pool can hold (Crew.max_proposals).
      required:
      - limit
      - proposals
      - size
...
    ProofKindEnum:
      enum:
      - none
      - photo
      - video
      - photo_or_video
      type: string
      description: |-
        * `none` - No proof
        * `photo` - Photo
        * `video` - Video
        * `photo_or_video` - Photo or video
...
    ScheduleInRequest:
      type: object
      properties:
        period_kind:
          $ref: '#/components/schemas/PeriodKindEnum'
        period_start:
          type: string
          format: date
          description: First day of the period (the 1st for a month).
      required:
      - period_kind
      - period_start
...
    TargetScopeEnum:
      enum:
      - none
      - per_check_in
      - per_week
      - per_period
      type: string
      description: |-
        * `none` - No target
        * `per_check_in` - Each check-in
        * `per_week` - Each week
        * `per_period` - The whole period
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: crew_session